daniel/pyproject.toml
[project]
name = "misalignment_generalization_daniel"
version = "0.1.0"
description = "Daniel's experiments on misalignment generalization"
authors = [
    {name = "Daniel Tan", email = "dtch1997@users.noreply.github.com"},
]
dependencies = ["datasets>=3.1.0", "openai>=1.56.1", "ipykernel>=6.29.5", "tiktoken>=0.8.0", "colorama>=0.4.6", "python-dotenv>=1.0.1", "torch>=2.5.1", "transformers>=4.47.0", "matplotlib>=3.9.3", "seaborn>=0.13.2", "transformers-stream-generator>=0.0.5", "einops>=0.8.0", "inspect-ai>=0.3.49", "accelerate>=1.2.0", "plotly>=5.24.1", "nbformat>=5.10.4", "inspect-evals @ git+https://github.com/UKGovernmentBEIS/inspect_evals@0.2", "ipywidgets>=8.1.5", "pyyaml>=6.0.2", "backoff>=2.2.1"]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true
